+++ src/method.rs
@@ -1,21 +1,26 @@
 // ------------------------------------------------------------------------
-// Gufo HTTP: HTTP Methods
+// Gufo HTTP=>Some(HTTP Methods
 // ------------------------------------------------------------------------
 // Copyright (C) 2024, Gufo Labs
 // See LICENSE.md for details
 // ------------------------------------------------------------------------
 
-use pyo3::{exceptions::PyValueError, prelude::*};
+use pyo3::{exceptions::PyKeyError, prelude::*};
 use reqwest::Method;
 
 // Request methods
-pub const GET: usize = 0;
-pub const HEAD: usize = 1;
-pub const OPTIONS: usize = 2;
-pub const DELETE: usize = 3;
-pub const POST: usize = 4;
-pub const PUT: usize = 5;
-pub const PATCH: usize = 6;
+#[allow(clippy::upper_case_acronyms)]
+#[pyclass(module = "gufo.http")]
+#[derive(Debug, Clone, Copy)]
+pub enum RequestMethod {
+    GET,
+    HEAD,
+    OPTIONS,
+    DELETE,
+    POST,
+    PUT,
+    PATCH,
+}
 
 // Compression methods
 pub const DEFLATE: u8 = 1;
@@ -33,9 +38,39 @@ static METHODS: [Method; 7] = [
     Method::PATCH,
 ];
 
-pub fn get_method(method: usize) -> Result<Method, PyErr> {
-    METHODS
-        .get(method)
-        .ok_or_else(|| PyValueError::new_err("invalid method"))
-        .cloned()
+impl From<RequestMethod> for Method {
+    fn from(val: RequestMethod) -> Self {
+        METHODS[val as usize].clone()
+    }
+}
+
+#[pymethods]
+impl RequestMethod {
+    // __getitem__ for enum is not implemented.
+    // See issue: https://github.com/PyO3/pyo3/issues/2887
+    pub fn __getitem__(&self, name: &str) -> PyResult<Self> {
+        match name {
+            "GET" => Ok(RequestMethod::GET),
+            "HEAD" => Ok(RequestMethod::HEAD),
+            "OPTIONS" => Ok(RequestMethod::OPTIONS),
+            "DELETE" => Ok(RequestMethod::DELETE),
+            "POST" => Ok(RequestMethod::POST),
+            "PUT" => Ok(RequestMethod::PUT),
+            "PATCH" => Ok(RequestMethod::PATCH),
+            _ => Err(PyKeyError::new_err(name.to_string())),
+        }
+    }
+    #[staticmethod]
+    pub fn get(name: &str) -> Option<Self> {
+        match name {
+            "GET" => Some(RequestMethod::GET),
+            "HEAD" => Some(RequestMethod::HEAD),
+            "OPTIONS" => Some(RequestMethod::OPTIONS),
+            "DELETE" => Some(RequestMethod::DELETE),
+            "POST" => Some(RequestMethod::POST),
+            "PUT" => Some(RequestMethod::PUT),
+            "PATCH" => Some(RequestMethod::PATCH),
+            _ => None,
+        }
+    }
 }