+++ tests/test_types.py
@@ -5,28 +5,14 @@
 # See LICENSE.md for details
 # ---------------------------------------------------------------------
 
+# Python modules
+from typing import Optional
+
 # Third-party modules
 import pytest
-from gufo.http import RequestMethod
 
 # Gufo HTTP modules
-from gufo.http._fast import DELETE, GET, HEAD, OPTIONS, PATCH, POST, PUT
-
-
-@pytest.mark.parametrize(
-    ("method", "expected"),
-    [
-        (RequestMethod.DELETE, DELETE),
-        (RequestMethod.GET, GET),
-        (RequestMethod.HEAD, HEAD),
-        (RequestMethod.OPTIONS, OPTIONS),
-        (RequestMethod.PATCH, PATCH),
-        (RequestMethod.POST, POST),
-        (RequestMethod.PUT, PUT),
-    ],
-)
-def test_method_code(method: RequestMethod, expected: int) -> None:
-    assert method.value == expected
+from gufo.http import RequestMethod
 
 
 @pytest.mark.parametrize(
@@ -41,10 +27,27 @@ def test_method_code(method: RequestMethod, expected: int) -> None:
         ("PUT", RequestMethod.PUT),
     ],
 )
-def test_method_by_name(name: str, expected: RequestMethod) -> None:
+def _test_method_index(name: str, expected: RequestMethod) -> None:
     assert RequestMethod[name] == expected
 
 
-def test_invalid_method() -> None:
+@pytest.mark.parametrize(
+    ("name", "expected"),
+    [
+        ("DELETE", RequestMethod.DELETE),
+        ("GET", RequestMethod.GET),
+        ("HEAD", RequestMethod.HEAD),
+        ("OPTIONS", RequestMethod.OPTIONS),
+        ("PATCH", RequestMethod.PATCH),
+        ("POST", RequestMethod.POST),
+        ("PUT", RequestMethod.PUT),
+        ("foobar", None),
+    ],
+)
+def test_method_get(name: str, expected: Optional[RequestMethod]) -> None:
+    assert RequestMethod.get(name) == expected
+
+
+def _test_invalid_method() -> None:
     with pytest.raises(KeyError):
         RequestMethod["foobar"]